# Conflicts:
#	commons/src/commonMain/kotlin/org/jetbrains/letsPlot/commons/intern/util/TextWidthEstimator.kt
#	demo/common-plot/src/commonMain/kotlin/demo/plot/common/model/EmptyGeomContext.kt
#	docs/dev/notebooks/latex.ipynb
#	plot-base/src/commonMain/kotlin/org/jetbrains/letsPlot/core/plot/base/aes/AesInitValue.kt
#	plot-base/src/commonMain/kotlin/org/jetbrains/letsPlot/core/plot/base/geom/TextGeom.kt
#	plot-base/src/commonMain/kotlin/org/jetbrains/letsPlot/core/plot/base/geom/util/LabelOptions.kt
#	plot-base/src/commonMain/kotlin/org/jetbrains/letsPlot/core/plot/base/geom/util/TextUtil.kt
#	plot-base/src/commonMain/kotlin/org/jetbrains/letsPlot/core/plot/base/render/svg/MultilineLabel.kt
#	plot-base/src/commonMain/kotlin/org/jetbrains/letsPlot/core/plot/base/render/svg/TextLabel.kt
#	plot-base/src/commonMain/kotlin/org/jetbrains/letsPlot/core/plot/base/render/text/Hyperlink.kt
#	plot-base/src/commonMain/kotlin/org/jetbrains/letsPlot/core/plot/base/render/text/Latex.kt
#	plot-base/src/commonMain/kotlin/org/jetbrains/letsPlot/core/plot/base/render/text/RichText.kt
#	plot-base/src/jvmTest/kotlin/org/jetbrains/letsPlot/core/plot/base/render/svg/MarkdownTest.kt
#	plot-base/src/jvmTest/kotlin/org/jetbrains/letsPlot/core/plot/base/render/svg/RichTextLatexTest.kt
#	plot-base/src/jvmTest/kotlin/org/jetbrains/letsPlot/core/plot/base/render/svg/RichTextTermTest.kt
#	plot-base/src/jvmTest/kotlin/org/jetbrains/letsPlot/core/plot/base/render/svg/TestUtil.kt
#	plot-builder/src/commonMain/kotlin/org/jetbrains/letsPlot/core/plot/builder/PlotSvgComponentHelper.kt
#	plot-builder/src/commonMain/kotlin/org/jetbrains/letsPlot/core/plot/builder/PlotTile.kt
#	plot-builder/src/commonMain/kotlin/org/jetbrains/letsPlot/core/plot/builder/assemble/GeomLayerBuilder.kt
#	plot-builder/src/commonMain/kotlin/org/jetbrains/letsPlot/core/plot/builder/defaultTheme/DefaultTheme.kt
#	plot-builder/src/commonMain/kotlin/org/jetbrains/letsPlot/core/plot/builder/defaultTheme/ThemeUtil.kt
#	plot-builder/src/commonMain/kotlin/org/jetbrains/letsPlot/core/plot/builder/guide/AxisComponent.kt
#	plot-builder/src/commonMain/kotlin/org/jetbrains/letsPlot/core/plot/builder/guide/ColorBarComponent.kt
#	plot-builder/src/commonMain/kotlin/org/jetbrains/letsPlot/core/plot/builder/guide/LegendBox.kt
#	plot-builder/src/commonMain/kotlin/org/jetbrains/letsPlot/core/plot/builder/guide/LegendComponent.kt
#	plot-builder/src/commonMain/kotlin/org/jetbrains/letsPlot/core/plot/builder/layout/PlotLayoutUtil.kt
#	plot-builder/src/commonMain/kotlin/org/jetbrains/letsPlot/core/plot/builder/presentation/PlotLabelSpec.kt
#	plot-builder/src/commonMain/kotlin/org/jetbrains/letsPlot/core/plot/builder/tooltip/component/TooltipBox.kt
#	plot-builder/src/jvmTest/kotlin/org/jetbrains/letsPlot/core/plot/builder/defaultTheme/ThemeColorOptionsTest.kt
#	plot-builder/src/jvmTest/kotlin/org/jetbrains/letsPlot/core/plot/builder/defaultTheme/values/ThemeOptionTest.kt
#	plot-builder/src/jvmTest/kotlin/org/jetbrains/letsPlot/core/plot/builder/layout/tile/PolarTileLayoutTest.kt
#	plot-builder/src/jvmTest/kotlin/org/jetbrains/letsPlot/core/plot/builder/layout/tile/TopDownTileLayoutTest.kt
#	plot-stem/src/jvmTest/kotlin/org/jetbrains/letsPlot/core/spec/config/GeomInteractionBuilderCreationTest.kt

+++ demo/common-plot/src/commonMain/kotlin/demo/plot/common/model/plotConfig/PolarHeatmap.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024. JetBrains s.r.o.
+ * Copyright (c) 2026. JetBrains s.r.o.
  * Use of this source code is governed by the MIT license that can be found in the LICENSE file.
  */
 
@@ -45,7 +45,7 @@ class PolarHeatmap {
 
         val (gridR, gridTheta) = simpleMeshgrid(
             linspace(0.0, maxR.toDouble(), stepsR),
-            linspace(0.0, maxTheta.toDouble(), stepsTheta)
+            linspace(0.0, maxTheta, stepsTheta)
         )
 
         val dataMap = getData(gridTheta.flatten(), gridR.flatten(), dataFunction)
